folder-gateway-skill/openapi_tools/requirements.txt
streamlit==1.40.2
gradio==4.44.1
httpx==0.28.1
PyYAML==6.0.3
